requirements.txt
altair
pandas
streamlit~=1.35.0

requests~=2.32.3
pillow~=10.3.0